feat: implement read-only Gen I and Gen II save parsers with StatExp and DV models

+++ src/save/src/CompanionReidentifier.cpp
@@ -10,7 +10,7 @@ Pocket::Companion::CompanionLink CompanionReidentifier::createLinkFromCreature(
 ) {
     Pocket::Companion::CompanionLink link;
     link.gameId = gameId;
-    link.generation = creature.generation;
+    link.generation = static_cast<int>(creature.generation);
     link.lastVerifiedSaveHash = saveHash;
     link.status = Pocket::Companion::LinkStatus::Linked;
